# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MoodBite API",
    "description": "Gợi ý nhà hàng theo ngữ cảnh: câu tìm kiếm tự do + vị trí + thời điểm.\n\nMọi response bọc trong `data` (thành công) hoặc `error` (lỗi).\nXem GET /api/v1/health để biết nguồn dữ liệu nào đã sẵn sàng.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "description": "Tạo tài khoản mới. Vai luôn là `user`.\n\n201 CREATED chứ không phải 200: có tài nguyên mới được tạo ra.\nTên đã có người dùng -> 409 USERNAME_TAKEN. Sai định dạng -> 400 INVALID_REQUEST.",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "description": "Đổi tài khoản/mật khẩu lấy token.\n\nSai tên HAY sai mật khẩu đều trả CÙNG MỘT câu 401 — không nói cái nào sai, nếu không\nthì đây thành công cụ dò xem tên nào đã tồn tại.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Forgot Password",
        "description": "Gửi thư kèm đường dẫn đặt lại mật khẩu.\n\n⚠️ LUÔN TRẢ CÙNG MỘT CÂU, dù tài khoản có tồn tại hay không, dù có email hay không.\nTrả lời khác nhau là biến endpoint này thành công cụ dò xem email/tên nào đã đăng ký —\nđúng thứ mà `/login` đã cẩn thận tránh.\n\nGiới hạn tần suất CHẶT hơn đăng nhập: mỗi lần gọi là một lá thư thật bay đi, và hạn\nmức SMTP miễn phí của Gmail chỉ khoảng 500 thư/ngày. Không chặn thì một người bấm liên\ntục là đủ làm tính năng chết cả ngày cho mọi người.",
        "operationId": "forgot_password_api_v1_auth_forgot_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Reset Password",
        "description": "Đổi mật khẩu bằng token trong thư.\n\nToken hỏng/hết hạn/đã dùng -> 401. Mật khẩu mới sai định dạng -> 400.\nĐổi xong KHÔNG tự đăng nhập: client phải chuyển sang trang đăng nhập.",
        "operationId": "reset_password_api_v1_auth_reset_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/change-password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Change Password",
        "description": "Đổi mật khẩu khi ĐANG đăng nhập. Khác `/reset-password` (dùng token trong thư).\n\nVẫn phải nhập mật khẩu HIỆN TẠI dù đã có token — xem `ChangePasswordUseCase`.\nGiới hạn tần suất dùng chung bộ đếm với đăng nhập: đây cũng là chỗ đoán mật khẩu được.",
        "operationId": "change_password_api_v1_auth_change_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangePasswordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/verify-email/confirm": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Confirm Email Verification",
        "description": "Đóng dấu đã xác minh bằng token trong thư. KHÔNG cần đăng nhập.\n\nCố ý không đòi đăng nhập: người dùng hay mở thư ở máy/trình duyệt khác, bắt đăng nhập\ntrước sẽ chặn đúng lúc họ vừa bấm vào link. Token đã tự mang chữ ký và thời hạn nên\nbản thân nó là bằng chứng đủ.\n\nToken hỏng/hết hạn/đã dùng/email đã đổi -> 401.\nDùng chung bộ đếm với đăng nhập vì đây cũng là chỗ có thể đem token ra thử.",
        "operationId": "confirm_email_verification_api_v1_auth_verify_email_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthData": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserSelf"
          },
          "token": {
            "type": "string",
            "title": "Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "description": "Số giây token còn hiệu lực"
          }
        },
        "type": "object",
        "required": [
          "user",
          "token",
          "expires_in"
        ],
        "title": "AuthData"
      },
      "AuthResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AuthData"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "AuthResponse"
      },
...
      "ChangePasswordData": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "token": {
            "type": "string",
            "title": "Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          }
        },
        "type": "object",
        "required": [
          "message",
          "token",
          "expires_in"
        ],
        "title": "ChangePasswordData",
        "description": "Đổi mật khẩu thu hồi MỌI token (kể cả token của máy đang gọi), nên phải trả kèm\ntoken mới cho máy này - thiếu nó thì người vừa đổi mật khẩu bị đá ra ngay."
      },
      "ChangePasswordRequest": {
        "properties": {
          "current_password": {
            "type": "string",
            "minLength": 1,
            "title": "Current Password"
          },
          "new_password": {
            "type": "string",
            "minLength": 1,
            "title": "New Password"
          }
        },
        "type": "object",
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "ChangePasswordRequest",
        "description": "Đổi mật khẩu khi ĐANG đăng nhập.\n\nVẫn đòi mật khẩu hiện tại dù đã có token — xem `ChangePasswordUseCase`."
      },
      "ChangePasswordResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ChangePasswordData"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ChangePasswordResponse"
      },
...
      "ForgotPasswordRequest": {
        "properties": {
          "identifier": {
            "type": "string",
            "maxLength": 254,
            "minLength": 1,
            "title": "Identifier",
            "description": "Email HOẶC tên đăng nhập. Người dùng thường chỉ nhớ một trong hai."
          }
        },
        "type": "object",
        "required": [
          "identifier"
        ],
        "title": "ForgotPasswordRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "MeResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UserSelf"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "MeResponse"
      },
      "MessageData": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Câu thông báo hiển thị thẳng cho người dùng."
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageData",
        "description": "Kết quả cho các thao tác không trả về tài nguyên nào."
      },
      "MessageResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MessageData"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "MessageResponse"
      },
...
      "RegisterRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Username",
            "description": "3-32 ký tự, chỉ chữ thường không dấu (a-z), số, gạch dưới, gạch ngang. Tự chuẩn hoá về chữ thường."
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "description": "Tối thiểu 8 ký tự. Không bắt buộc hoa/số/ký tự đặc biệt — độ dài quan trọng hơn (NIST SP 800-63B)."
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name",
            "description": "Tên hiển thị, ĐƯỢC dùng tiếng Việt có dấu. Không dùng để đăng nhập."
          },
          "email": {
            "type": "string",
            "maxLength": 254,
            "minLength": 1,
            "title": "Email",
            "description": "BẮT BUỘC (đổi 2026-08-24). Dùng để xác minh tài khoản và lấy lại mật khẩu. Trước đó là tuỳ chọn; chủ dự án đổi sau khi luồng xác minh email hoàn thành."
          }
        },
        "type": "object",
        "required": [
          "username",
          "password",
          "email"
        ],
        "title": "RegisterRequest"
      },
      "ResetPasswordRequest": {
        "properties": {
          "token": {
            "type": "string",
            "minLength": 1,
            "title": "Token",
            "description": "Token lấy từ đường dẫn trong thư."
          },
          "new_password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "New Password",
            "description": "Mật khẩu mới, tối thiểu 8 ký tự."
          }
        },
        "type": "object",
        "required": [
          "token",
          "new_password"
        ],
        "title": "ResetPasswordRequest"
      },
...
      "UserSelf": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "user | admin"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "Chỉ chính chủ thấy. Không lộ cho người khác."
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At",
            "description": "Ngày tạo tài khoản, dạng ISO-8601."
          },
          "email_verified": {
            "type": "boolean",
            "title": "Email Verified",
            "description": "Hộp thư đã được chứng minh là có thật chưa (chủ tài khoản đã bấm đường dẫn trong thư). Chưa xác minh KHÔNG chặn đăng nhập — email vốn là trường tuỳ chọn.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "username",
          "role"
        ],
        "title": "UserSelf",
        "description": "Bản hồ sơ CHÍNH CHỦ nhìn thấy — có thêm email và ngày tham gia.\n\nKế thừa UserPublic để hai bên không lệch nhau khi thêm trường mới."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VerifyEmailRequest": {
        "properties": {
          "token": {
            "type": "string",
            "minLength": 1,
            "title": "Token",
            "description": "Token lấy từ đường dẫn trong thư."
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "VerifyEmailRequest",
        "description": "Xác minh email bằng token trong thư.\n\nKHÔNG có trường email: địa chỉ nằm sẵn trong token và đã được ký. Nhận email từ\nclient rồi tin theo là mở đường cho người ta xác minh hộ địa chỉ của người khác."
      },
...
    }
  }
}