# src/presentation/api/routers/auth.py
"""Router tài khoản người dùng — `/api/v1/auth/*`.

MỎNG theo đúng CLAUDE.md mục 3: nhận HTTP, gọi use case, trả envelope. Quy tắc đặt tên
và mật khẩu nằm ở `domain/entities/user.py`; băm và ký token nằm ở `infrastructure/auth/`.

QUÊN MẬT KHẨU (thêm 2026-08-22): `/forgot-password` gửi thư, `/reset-password` đổi mật
khẩu bằng token trong thư. Token KHÔNG lưu vào CSDL — xem `infrastructure/auth/password_reset.py`.

BA CHỐT CHẶN, độc lập nhau:
  1. Chưa đặt MOODBITE_AUTH_SECRET -> 503 kèm hướng dẫn. Fail-closed.
  2. Giới hạn tần suất theo IP trên cả `/register` lẫn `/login`.
  3. Vai LUÔN là `user`. Router KHÔNG nhận `role` từ client - xem `RegisterUserUseCase`.

`/logout` THU HỒI THẬT (từ 2026-09-29): token mang `tv`, bảng `users` có `token_version`;
đăng xuất / đổi mật khẩu / đặt lại mật khẩu tăng số đó nên mọi token cũ hết giá trị ngay.
Trước đó dự án cố ý KHÔNG có `/logout`, vì một endpoint trả 200 mà không thu hồi được gì
là ảo giác an toàn - xem docs/API_DECISIONS_PENDING.md §5.8.
"""
from __future__ import annotations

from fastapi import APIRouter, Depends, Request

from src.domain.entities.user import User
from src.presentation.api.dependencies import (
    Container,
    client_key,
    get_container,
    get_current_user,
)
from src.presentation.api.envelope import success
from src.presentation.api.schemas import (
    AuthResponse,
    ChangePasswordRequest,
    ChangePasswordResponse,
    ForgotPasswordRequest,
    LoginRequest,
    MeResponse,
    MessageResponse,
    RegisterRequest,
    ResetPasswordRequest,
    VerifyEmailRequest,
)

router = APIRouter(prefix="/auth", tags=["auth"])


def _auth_payload(user: User, token: str, ttl_seconds: int) -> dict:
    """Hình dạng response dùng chung cho cả đăng ký và đăng nhập.

    Đăng ký trả luôn token để người dùng không phải đăng nhập lại ngay sau đó — một bước
    thừa mà ai cũng bỏ qua được.

    `to_self()` (không phải `to_public()`): người nhận vừa chứng minh xong họ là chủ tài
    khoản — gõ đúng mật khẩu, hoặc vừa tự tạo tài khoản này. Nhờ vậy frontend biết ngay
    `email_verified` mà không phải gọi thêm một vòng `/auth/me`.

    Entity là nơi DUY NHẤT quyết định trường nào được lộ ra; router không tự dựng dict,
    nếu không thì thêm trường vào entity sẽ có ngày lộ cả `password_hash`.
    """
    return {
        "user": user.to_self(),
        "token": token,
        "token_type": "bearer",
        "expires_in": ttl_seconds,
    }


@router.post("/register", response_model=AuthResponse, status_code=201)
def register(
    payload: RegisterRequest,
    request: Request,
    container: Container = Depends(get_container),
):
    """Tạo tài khoản mới. Vai luôn là `user`.

    201 CREATED chứ không phải 200: có tài nguyên mới được tạo ra.
    Tên đã có người dùng -> 409 USERNAME_TAKEN. Sai định dạng -> 400 INVALID_REQUEST.
    """
    container.user_tokens.ensure_configured()
    # Đếm TRƯỚC khi băm mật khẩu: băm tốn ~0.4s CPU, để sau thì kẻ tấn công vẫn ép được
    # server làm việc nặng dù request rốt cuộc bị từ chối.
    container.register_rate_limiter.check(client_key(request))

    user, token = container.register_user.execute(
        payload.username, payload.password, payload.display_name, payload.email
    )

    # Có khai email thì gửi luôn thư xác minh — người dùng không phải tìm nút bấm.
    # `try_send` NUỐT lỗi gửi thư, cố ý: tài khoản đã ghi vào CSDL rồi, để lỗi SMTP ném
    # ra đây thì người dùng thấy "đăng ký thất bại" trong khi tên đã bị chiếm và họ không
    # đăng ký lại được nữa. Xem `RequestEmailVerificationUseCase.try_send`.
    container.request_email_verification.try_send(user)

    return success(
        _auth_payload(user, token, container.user_tokens.token_ttl_seconds),
        status_code=201,
    )


@router.post("/login", response_model=AuthResponse)
def login(
    payload: LoginRequest,
    request: Request,
    container: Container = Depends(get_container),
):
    """Đổi tài khoản/mật khẩu lấy token.

    Sai tên HAY sai mật khẩu đều trả CÙNG MỘT câu 401 — không nói cái nào sai, nếu không
    thì đây thành công cụ dò xem tên nào đã tồn tại.
    """
    container.user_tokens.ensure_configured()
    key = client_key(request)
    container.login_rate_limiter.check(key)

    user, token = container.login_user.execute(payload.username, payload.password)

    # Đăng nhập ĐÚNG thì xoá lịch sử đếm: người gõ nhầm vài lần rồi vào được không đáng
    # bị chặn oan ở lần đăng nhập sau.
    container.login_rate_limiter.reset(key)
    return success(_auth_payload(user, token, container.user_tokens.token_ttl_seconds))


@router.post("/forgot-password", response_model=MessageResponse)
def forgot_password(
    payload: ForgotPasswordRequest,
    request: Request,
    container: Container = Depends(get_container),
):
    """Gửi thư kèm đường dẫn đặt lại mật khẩu.

    ⚠️ LUÔN TRẢ CÙNG MỘT CÂU, dù tài khoản có tồn tại hay không, dù có email hay không.
    Trả lời khác nhau là biến endpoint này thành công cụ dò xem email/tên nào đã đăng ký —
    đúng thứ mà `/login` đã cẩn thận tránh.

    Giới hạn tần suất CHẶT hơn đăng nhập: mỗi lần gọi là một lá thư thật bay đi, và hạn
    mức SMTP miễn phí của Gmail chỉ khoảng 500 thư/ngày. Không chặn thì một người bấm liên
    tục là đủ làm tính năng chết cả ngày cho mọi người.
    """
    container.reset_tokens.ensure_configured()
    container.forgot_password_rate_limiter.check(client_key(request))

    # Lỗi máy chủ thư thì VẪN ném ra ngoài (503): đó không phải thông tin về tài khoản,
    # và giấu đi thì người dùng ngồi đợi mãi một lá thư không bao giờ tới.
    container.request_password_reset.execute(payload.identifier)

    return success(
        {
            "message": (
                "Nếu tài khoản tồn tại và đã khai email, thư hướng dẫn đặt lại mật khẩu "
                "đã được gửi. Hãy kiểm tra cả hộp thư rác."
            )
        }
    )


@router.post("/reset-password", response_model=MessageResponse)
def reset_password(
    payload: ResetPasswordRequest,
    request: Request,
    container: Container = Depends(get_container),
):
    """Đổi mật khẩu bằng token trong thư.

    Token hỏng/hết hạn/đã dùng -> 401. Mật khẩu mới sai định dạng -> 400.
    Đổi xong KHÔNG tự đăng nhập: client phải chuyển sang trang đăng nhập.
    """
    container.reset_tokens.ensure_configured()
    container.login_rate_limiter.check(client_key(request))

    container.reset_password.execute(payload.token, payload.new_password)
    return success({"message": "Đã đổi mật khẩu. Hãy đăng nhập bằng mật khẩu mới."})


@router.post("/change-password", response_model=ChangePasswordResponse)
def change_password(
    payload: ChangePasswordRequest,
    request: Request,
    user: User = Depends(get_current_user),
    container: Container = Depends(get_container),
):
    """Đổi mật khẩu khi ĐANG đăng nhập. Khác `/reset-password` (dùng token trong thư).

    Vẫn phải nhập mật khẩu HIỆN TẠI dù đã có token — xem `ChangePasswordUseCase`.
    Giới hạn tần suất dùng chung bộ đếm với đăng nhập: đây cũng là chỗ đoán mật khẩu được.
    """
    container.login_rate_limiter.check(client_key(request))
    cap_nhat = container.change_password.execute(
        user, payload.current_password, payload.new_password
    )
    return success(
        {
            "message": "Đã đổi mật khẩu. Các thiết bị khác đã bị đăng xuất.",
            # Token mới mang `token_version` mới - token đang dùng đã bị thu hồi cùng lúc.
            "token": container.user_tokens.issue(cap_nhat),
            "token_type": "bearer",
            "expires_in": container.user_tokens.token_ttl_seconds,
        }
    )


@router.post("/logout", response_model=MessageResponse)
def logout(
    user: User = Depends(get_current_user),
    container: Container = Depends(get_container),
):
    """Đăng xuất THẬT: thu hồi mọi token của tài khoản, ở MỌI thiết bị.

    Không đăng xuất riêng từng máy được - đánh đổi của phương án `token_version`, xem
    `LogoutEverywhereUseCase`. Câu trả lời nói rõ điều đó.
    """
    container.logout_everywhere.execute(user)
    return success({"message": "Đã đăng xuất trên mọi thiết bị."})


@router.post("/verify-email/request", response_model=MessageResponse)
def request_email_verification(
    request: Request,
    user: User = Depends(get_current_user),
    container: Container = Depends(get_container),
):
    """Gửi lại thư xác minh email cho tài khoản ĐANG đăng nhập.

    KHÁC `/forgot-password` ở chỗ KHÔNG phải giấu "tài khoản có tồn tại không": người gọi
    đã đăng nhập nên đã biết thừa tài khoản của chính họ. Vì vậy ở đây nói thẳng được
    "chưa khai email" hay "đã xác minh rồi" — giấu chỉ làm người dùng bối rối.

    Dùng CHUNG bộ đếm với `/forgot-password`: cả hai đều làm một lá thư thật bay đi, và
    hạn mức SMTP miễn phí của Gmail chỉ khoảng 500 thư/ngày.
    """
    container.email_verify_tokens.ensure_configured()
    container.forgot_password_rate_limiter.check(client_key(request))

    da_gui = container.request_email_verification.execute(user)
    if not da_gui:
        thong_diep = (
            "Email của bạn đã được xác minh rồi."
            if user.email_verified
            else "Tài khoản chưa khai email nên không gửi được thư xác minh."
        )
        return success({"message": thong_diep})
    return success(
        {"message": "Đã gửi thư xác minh. Hãy kiểm tra cả hộp thư rác."}
    )


@router.post("/verify-email/confirm", response_model=MeResponse)
def confirm_email_verification(
    payload: VerifyEmailRequest,
    request: Request,
    container: Container = Depends(get_container),
):
    """Đóng dấu đã xác minh bằng token trong thư. KHÔNG cần đăng nhập.

    Cố ý không đòi đăng nhập: người dùng hay mở thư ở máy/trình duyệt khác, bắt đăng nhập
    trước sẽ chặn đúng lúc họ vừa bấm vào link. Token đã tự mang chữ ký và thời hạn nên
    bản thân nó là bằng chứng đủ.

    Token hỏng/hết hạn/đã dùng/email đã đổi -> 401.
    Dùng chung bộ đếm với đăng nhập vì đây cũng là chỗ có thể đem token ra thử.
    """
    container.email_verify_tokens.ensure_configured()
    container.login_rate_limiter.check(client_key(request))

    user = container.confirm_email_verification.execute(payload.token)
    # `to_self()`: người vừa chứng minh mình cầm được thư gửi tới hộp thư của tài khoản
    # này, nên cho xem hồ sơ của chính tài khoản đó là hợp lý.
    return success(user.to_self())


@router.get("/me", response_model=MeResponse)
def me(user: User = Depends(get_current_user)):
    """Thông tin tài khoản đang đăng nhập, kèm VAI hiện tại.

    Frontend gọi endpoint này lúc mở app để biết token còn sống không và được vào những
    đâu. Vai đọc từ CSDL chứ không lấy trong token, nên admin vừa bị hạ quyền sẽ thấy
    ngay ở lần gọi kế tiếp.
    """
    # `to_self()` (không phải `to_public()`): đây là CHÍNH CHỦ xem hồ sơ mình, nên được
    # thấy email và ngày tham gia của chính mình. Xem ghi chú ở `domain/entities/user.py`.
    return success(user.to_self())